requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# hee-blog-post-migration/src/blog_extractor.py: 3
beautifulsoup4 == 4.9.3

# hee-blog-post-migration/src/blog_brxm_yaml_decorator.py: 5
# hee-blog-post-migration/src/main.py: 3
ruamel.yaml == 0.17.4

# hee-blog-post-migration/src/blog_category_util.py: 2
sortedcontainers == 2.3.0
